+++ libopenspotify/playlist.h
@@ -17,7 +17,7 @@ sp_playlist *playlist_create(sp_session *session, unsigned char id[17]);
 void playlist_set_name(sp_session *session, sp_playlist *playlist, char *name);
 void playlist_release(sp_session *session, sp_playlist *playlist);
 
-unsigned long playlist_checksum(sp_playlist *playlist);
-unsigned long playlistcontainer_checksum(sp_playlistcontainer *container);
+unsigned int playlist_checksum(sp_playlist *playlist);
+unsigned int playlistcontainer_checksum(sp_playlistcontainer *container);
 
 #endif
